Initial PoC of firewalld zone for cluster internal communication.
This configure a set of zone, ipset and service files for firewalld in
/usr/lib/firewalld, allowing for local customizations using firewall-cmd,
which will then store changes in /etc/firewalld.

I don't yet see how to figure out the daemon IP addresses of the cluster
nodes, so these needs to be added manually to the ipset on all nodes for now:

    # firewall-cmd --permanent --ipset=daemon-network --add-entry=10.11.37.77
    # firewall-cmd --permanent --ipset=daemon-network --add-entry=10.11.37.86
    # firewall-cmd --permanent --ipset=daemon-network --add-entry=10.11.37.173
    # firewall-cmd --permanent --ipset=daemon-network --get-entries
    10.11.37.86
    10.11.37.77
    10.11.37.173

It's assuming tscCmdPortRange=60000-61000, but that should be easy enough to set
more dynamically.

Signed-off-by: Jan-Frode Myklebust <janfrode@tanso.net>